requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
chardet==3.0.4
colorama==0.4.6
distro==1.9.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==5.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pydantic==2.7.1
pydantic_core==2.18.2
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
pytesseract==0.3.10
python-docx==0.8.11
pywin32-ctypes==0.2.2
rfc3986==1.5.0
setuptools==69.5.1
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
